+++ src/lib/core/importer.ts
@@ -1,6 +1,6 @@
 import { glob } from 'glob';
 import { resolve } from 'path';
-import { AUTO_INJECTABLE_WATERMARK } from '../constants';
+import { AUTO_INJECTABLE_WATERMARK } from '../interfaces/auto-injectable.contant';
 
 type ClassType = new (...args: any[]) => any;
