+++ src/lib/decorators/component-scan.decorator.ts
