+++ src/lib/decorators/index.ts
@@ -1,0 +1,2 @@
+export * from './auto-injectable.decorator';
+export * from './component-scan.decorator';
