+++ src/lib/index.ts
@@ -1,0 +1,4 @@
+export * from './interfaces';
+export * from './core';
+export * from './decorators';
+export * from './auto.module';
