+++ src/lib/interfaces/auto-injectable.contant.ts
