+++ test/fixture.ts
@@ -1,4 +1,4 @@
-import { AutoInjectable } from '../src/lib/decorator/auto-injectable.decorator';
+import { AutoInjectable } from '../src/lib/decorators/auto-injectable.decorator';
 
 @AutoInjectable()
 export class AutoInjectableFixture {}
